Update README.md
Study Sesh is a CRUD web application. This Web application allows for users to make an account and authenticates information on login. 
Once a user is logged in they can create Flash Card Decks and populate each deck with unique cards that have a question, and answer, and an optional hint. 
After creating a deck, the user can proceed to study the deck. The application gives each user 3 attempts on incorrect answers and logs correct/incorrect questions. 
The app will then display the logged information on a review page after the test in completed.